refactor(web): extract shared collectFiles helper for TV build scripts
